requirements.txt
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.8
django-contact-form==1.4.2
django-modelcluster==3.1
django-registration==2.2
django-taggit==0.22.1
django-treebeard==4.2.0
djangorestframework==3.6.4
gunicorn==19.7.1
html5lib==0.999999999
idna==2.6
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
six==1.11.0
static3==0.7.0
Unidecode==0.4.21
urllib3==1.22
wagtail==1.13.1
webencodings==0.5.1
whitenoise==3.3.1
Willow==1.1
